% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{colour_picker}
\alias{color_picker}
\alias{colour_picker}
\title{A simple colour picker}
\usage{
colour_picker(hcl = TRUE)

color_picker(hcl = TRUE)
}
\arguments{
\item{hcl}{If \code{TRUE} (default), transforms the colour names
to hcl using \code{scale::col2hcl}.}
}
\value{
A vector of colours.
}
\description{
A simple interface to manually pick colours from
\code{colours()}. Uses the current device is none is open.
}
\author{
Laurent Gatto
}
\seealso{
The \code{codourpicker} package by Dean Attali is a much
    more elaborate version of this function
    \url{https://github.com/daattali/colourpicker}.
}

