% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cputools}
\alias{cputools}
\title{Available cputools functions}
\usage{
cputools()
}
\value{
Used for its side effect of print a list of available
helper functions.
}
\description{
Print all available (exported) cputools functions.
}
\examples{
cputools()
}
\author{
Laurent Gatto
}

