% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpu.R
\name{cpuwww}
\alias{cpuwww}
\title{Return the CPU webpage url.}
\usage{
cpuwww()
}
\value{
The URL as a \code{character}
}
\description{
Return the CPU webpage url.
}
\examples{
cpuwww()
}
\author{
Laurent Gatto
}

