% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{darken}
\alias{darken}
\alias{lighten}
\title{darken and lighten colours}
\usage{
darken(color, factor = 1.4)

lighten(color, factor = 1.4)
}
\arguments{
\item{color}{Colour name.}

\item{factor}{Lightening/darkening factory.}
}
\value{
A new rgb colour
}
\description{
Light and dark colours
}
\details{
From \url{https://gist.github.com/Jfortin1/72ef064469d1703c6b30}.
}
\examples{
darkFirebrick <- darken("firebrick")
lightFirebrick <- lighten("firebrick")
plot(c(1.5, 2, 2.5), c(1, 1.25, 1),
     pch = 19, cex = 150,
     col = c(darkFirebrick, "firebrick", lightFirebrick))
}

