% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\name{ht}
\alias{ht}
\title{head and tail}
\usage{
ht(x, ...)
}
\arguments{
\item{x}{An object for which  \code{head}/\code{tail} are defined.}

\item{...}{Additional parameters passed to \code{head} and \code{tail}.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Look at head and tail in one go.
}
\examples{
ht(letters)
ht(freeny.x, n = 3)
}
\author{
Laurent Gatto
}

