% Generated by roxygen2 (4.0.0): do not edit by hand
\name{makeFigName}
\alias{makeFigName}
\title{Create a file name for a figure}
\usage{
makeFigName(prefix, ext = "pdf", prefix0 = "Fig-", date = TRUE, path, ...)
}
\arguments{
\item{prefix}{A file specific prefix.}

\item{ext}{The extension; default is \code{pdf}.}

\item{prefix0}{The general prefix for figures. Default is \code{Fig}.}

\item{date}{Should the data (as reported by \code{date}) be used
as part of the file name? Default is \code{TRUE}.}

\item{path}{An optional path to the figure file.}

\item{...}{Currently ignored.}
}
\value{
A \code{character} with the figure file name.
}
\description{
Creates a figure file name given a prefix and other optional
paramters. Checks if file exists.
}
\details{
TODO: If file exists, create the next increment.
}
\examples{
makeFigName("foo")
makeFigName("foo", date = FALSE)
makeFigName("foo", path = "~/projects/big-project/figs")
}
\author{
Laurent Gatto
}

