% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/file-names.R
\name{pdfpng}
\alias{pdfpng}
\title{Plot to pdf and png}
\usage{
pdfpng(expr, fn)
}
\arguments{
\item{expr}{A quoted expression that produces the desired plot.}

\item{fn}{The file name (without extension).}
}
\value{
Produces the files \code{fn.pdf} and \code{fn.png}.
}
\description{
Plot to the pdf and png devices at once.
}
\examples{
p <- quote({
   plot(1:10)
   points(10:1)
})
pdfpng(p, "myfig")
}
\author{
Laurent Gatto
}

