% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shields.R
\name{makeTravisShield}
\alias{makeBiocBuildShield}
\alias{makeBiocCovrShield}
\alias{makeTravisShield}
\title{Generate the markdown code the a package shields.}
\usage{
makeTravisShield(pkg, user = options()[["GitHubUserName"]],
  branch = "master")

makeBiocBuildShield(pkg, branch = c("devel", "release"))

makeBiocCovrShield(pkg, branch = c("devel", "release"))
}
\arguments{
\item{pkg}{Mandatory \code{character} with the package name(s). If
of length greater than one, then must of of same length as
\code{user} and \code{branch}.}

\item{user}{Mandatory GitHub username. Default is to retrieve
\code{options()[["GitHubUserName"]]}. Argument is not recycled
of more than one package is provided.}

\item{branch}{Mandatory branch name. Default is
\code{"master"}. Argument is not recycled of more than one
package is provided. For Bioconductor shields (functions
\code{makeBiocBuildShield} and \code{makeBiocCovrShield}),
branch is one of \code{"devel"} (default) or \code{"release"}.}
}
\value{
A \code{character} with the shield in markdown.
}
\description{
Functions to generate package travis-ci, codecov, Bioconductor
build status and Biocondcutor test coverage shields
}
\examples{
makeTravisShield("pRoloc")
makeBiocBuildShield("MSnbase")
makeBiocBuildShield("MSnbase", branch = "release")
makeBiocCovrShield("MSnbase")
makeBiocCovrShield("MSnbase", branch = "release")
}
\author{
Laurent Gatto
}

