% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colours.R
\name{colour_picker}
\alias{color_picker}
\alias{colour_picker}
\title{A simple colour picker}
\usage{
colour_picker()

color_picker()
}
\value{
A vector of colours.
}
\description{
A simple interface to manually pick colours from
\code{colours()}. Uses the current device is none is open.
}
\author{
Laurent Gatto
}

