% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cputools.R
\name{cputools}
\alias{cputools}
\title{Available cputools functions}
\usage{
cputools()
}
\value{
Used for its side effect of print a list of available
helper functions.
}
\description{
Print all available (exported) cputools functions.
}
\examples{
cputools()
}
\author{
Laurent Gatto
}

