% Generated by roxygen2 (4.0.0): do not edit by hand
\name{makeFigName}
\alias{makeDatName}
\alias{makeFigName}
\title{Create a file name.}
\usage{
makeFigName(prefix, ext = "pdf", prefix0 = "Fig", date = TRUE,
  path = ".")

makeDatName(prefix, ext = "rda", prefix0 = "Dat", date = TRUE,
  path = ".")
}
\arguments{
\item{prefix}{A file specific prefix.}

\item{ext}{The extension; default is \code{pdf} for a figure and
\code{rda} for a data file.}

\item{prefix0}{The general prefix for figures. Default is
\code{Fig} for a figure and \code{Dat} for a data file.}

\item{date}{Should the data (as reported by \code{date}) be used
as part of the file name? Default is \code{TRUE}.}

\item{path}{The path to the file. Default is \code{"."}.}
}
\value{
A \code{character} with the figure file name.
}
\description{
Creates a file name given a prefix and other optional paramters
for a figure or data file. Also checks if file exists.
}
\details{
TODO: If file exists, create the next increment.
}
\examples{
makeFigName("foo")
makeFigName("foo", date = FALSE)
makeDatName("foo", date = FALSE)
makeFigName("foo", path = "~/projects/big-project/figs")
makeDatName("foo", path = "~/projects/big-project/Data")
}
\author{
Laurent Gatto
}

