% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{makeGithubUrl}
\alias{makeGithubIssuesUrl}
\alias{makeGithubPageUrl}
\alias{makeGithubUrl}
\alias{pkgqsts}
\title{Generates Github URLs}
\usage{
makeGithubUrl(pkg, user = options()[["GitHubUserName"]])

makeGithubPageUrl(pkg, user = options()[["GitHubUserName"]])

makeGithubIssuesUrl(pkg, user = options()[["GitHubUserName"]])

pkgqsts(pkg, github = TRUE, bioc = TRUE, level = 1L,
  user = options()[["GitHubUserName"]])
}
\arguments{
\item{pkg}{Mandatory \code{character} with the package name. Must
be of length one for code{pkgqsts}. Other functions
\code{makeGithubPageUrl} and \code{makeGithubIssuesUrl} are
vectorised.}

\item{user}{Mandatory GitHub username. Default is to retrieve
\code{options()[["GitHubUserName"]]}.}

\item{github}{A \code{logical} (default is \code{TRUE}) defining
whether users can ask questions on GitHub issues.}

\item{bioc}{A \code{logical} (default is \code{TRUE}) defining
whether users can ask questions on the Bioconductor support
forum.}

\item{level}{Level for markdown sections. Default is \code{1L} for
\code{#}. Use \code{2L} for \code{##} and so on.}
}
\value{
A \code{character} with Github URLs.

A \code{character} with Github page URLs.

A \code{character} with Github page URLs.

Invisibly returns \code{NULL}. Used for its side effect.
}
\description{
Generates a markdown section on where to ask questions about a
package.
}
\examples{
makeGithubUrl("pRoloc", user = "lgatto")
makeGithubUrl(c("pRoloc", "pRolocGUI"), user = "lgatto")
makeGithubPageUrl("pRoloc", user = "lgatto")
makeGithubPageUrl(c("pRoloc", "pRolocGUI"), user = "lgatto")
makeGithubIssuesUrl("pRoloc", user = "lgatto")
makeGithubIssuesUrl(c("pRoloc", "pRolocGUI"), user = "lgatto")
pkgqsts("pRoloc", user = "lgatto")
pkgqsts("cputools", bioc = FALSE, user = "lgatto")
}
\author{
Laurent Gatto
}
\seealso{
\code{BiocStyle::Githubpkg}
}

